requirements.txt
fastapi
httpx
beautifulsoup4
pydantic
python-dotenv
redis
